stoi/stof error

+++ src/Medicine.cpp
@@ -6,7 +6,7 @@ Medicine::Medicine() {
 
 }
 
-Medicine::Medicine(string n, string desc, float p, float q, float iva, int c, float disc, bool nr)
+Medicine::Medicine(string n, string desc, float p, int q, float iva, int c, float disc, bool nr)
 : Product(n, desc, p,q, iva, c, true), discount(disc), prescription(nr){}
